Established vue-router to switch in two pages.
