+++ index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <title>sellpage</title>
+    <title>SE Retail Management System</title>
   </head>
   <body>
     <div id="app"></div>
