+++ src/components/Storage.vue
@@ -1,0 +1,52 @@
+<template>
+  <div id="app">
+    <h1>You are in storage management page.</h1>
+  </div>
+</template>
+
+<style>
+  #app {
+    font-family: "Avenir", Helvetica, Arial, sans-serif;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    text-align: center;
+    color: #2c3e50;
+  }
+
+  header {
+    z-index: 1000;
+    min-width: 1200px;
+    transition: all 0.5s ease;
+    border-top: solid 4px #3091f2;
+    background-color: #fff;
+    box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.12), 0 0 6px 0 rgba(0, 0, 0, 0.04);
+  }
+
+  header.header-fixed {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+  }
+
+  main {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    min-height: 800px;
+    border: solid 40px #ffffff;
+    background-color: #ffffff;
+  }
+
+  main .el-card {
+    margin-bottom: 15px;
+  }
+
+  main.el-table {
+    text-align: center;
+  }
+
+  main.el-col {
+    margin: 10px;
+  }
+</style>
