+++ src/router/index.js
@@ -1,6 +1,5 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Hello from '@/components/Hello'
 
 Vue.use(Router)
 
@@ -8,8 +7,17 @@ export default new Router({
   routes: [
     {
       path: '/',
-      name: 'Hello',
-      component: Hello
+      component: resolve => require(['@/components/common/Home.vue'], resolve),
+      children: [
+        {
+          path: '/selling',
+          component: resolve => require(['@/components/Selling.vue'], resolve)
+        },
+        {
+          path: '/storage',
+          component: resolve => require(['@/components/Storage.vue'], resolve)
+        }
+      ]
     }
   ]
 })
